refactor: 도메인 엔티티만 테스트하기위해 클래스 분리
