+++ docs/QnA.md
@@ -50,7 +50,7 @@
    - 연관관계 편의메서드 작성
  - [X] DeleteHistory에서 시작하는 연관관계 매핑
  - [X] 도메인 엔티티 및 로직 분석
- - [ ] repository없이 도메인 엔티티만을 테스트할수 있도록 리팩토링
+ - [X] 도메인 엔티티만 테스트하기위해 클래스 분리
 
 
 # 리뷰 내용 반영
